backend/requirements.txt
# SpeakMate AI - Production Dependencies

# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Google Cloud
google-cloud-speech==2.24.1
google-cloud-texttospeech==2.16.3
google-generativeai==0.4.0

# Supabase
supabase>=2.0.0
python-jose[cryptography]==3.3.0

# Background Jobs (Redis + RQ)
redis==5.0.1
rq==1.16.0

# PDF Generation
reportlab==4.1.0
Pillow==10.2.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
python-dotenv==1.0.1
httpx>=0.24.0,<0.26.0
aiofiles==23.2.1
tenacity==8.2.3

# Audio processing
pydub==0.25.1
numpy>=1.24.0

# Rate limiting and security
slowapi==0.1.9

# Monitoring
sentry-sdk[fastapi]>=1.0.0
prometheus-client>=0.19.0
structlog>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# JSON Schema validation
jsonschema>=4.0.0

# Telegram Bot
aiogram>=3.4.0
